+++ bosesoundtouchapi/__init__.py
@@ -11,13 +11,11 @@ _________________
 """
 
 # our package imports.
-from bosesoundtouchapi.soundtouchaudiomodes import SoundTouchAudioModes
 from bosesoundtouchapi.soundtouchclient import SoundTouchClient
 from bosesoundtouchapi.soundtouchdevice import SoundTouchDevice
 from bosesoundtouchapi.soundtouchdiscovery import SoundTouchDiscovery
 from bosesoundtouchapi.soundtoucherror import SoundTouchError
 from bosesoundtouchapi.soundtouchexception import SoundTouchException
-from bosesoundtouchapi.soundtouchhdmicecmodes import SoundTouchHdmiCecModes
 from bosesoundtouchapi.soundtouchitemtypes import SoundTouchItemTypes
 from bosesoundtouchapi.soundtouchkeys import SoundTouchKeys
 from bosesoundtouchapi.soundtouchmessage import SoundTouchMessage
@@ -27,13 +25,11 @@ from bosesoundtouchapi.soundtouchwarning import SoundTouchWarning
 
 # all classes to import when "import *" is specified.
 __all__ = [
-    'SoundTouchAudioModes',
     'SoundTouchClient',
     'SoundTouchDevice',
     'SoundTouchDiscovery',
     'SoundTouchError',
     'SoundTouchException',
-    'SoundTouchHdmiCecModes',
     'SoundTouchItemTypes',
     'SoundTouchKeys',
     'SoundTouchMessage',