+++ bosesoundtouchapi/models/audiodspaudiomodes.py
@@ -2,36 +2,36 @@
 from enum import Enum
 
 # our package imports.
-from .bstutils import export
+from ..bstutils import export
 
 @export
-class SoundTouchAudioModes(Enum):
+class AudioDspAudioModes(Enum):
     """
-    SoundTouch Audio Modes.
+    Audio DSP Audio Modes enumeration.
     """
     
-    DIALOG = 'AUDIO_MODE_DIALOG'
+    Dialog = 'AUDIO_MODE_DIALOG'
     """
     Dialogue mode improves the clarity of dialogue and vocals in movies, 
     TV programs and podcasts by adjusting the tonal balance of the system.
     """
 
-    DIRECT = 'AUDIO_MODE_DIRECT'
+    Direct = 'AUDIO_MODE_DIRECT'
     """
     No optimizations made; the audio is directly from the source.
     """
 
-    NIGHT = 'AUDIO_MODE_NIGHT'
+    Night = 'AUDIO_MODE_NIGHT'
     """
     Audio Mode for a nighttime environment.
     """
 
-    NORMAL = 'AUDIO_MODE_NORMAL'
+    Normal = 'AUDIO_MODE_NORMAL'
     """
     Audio Mode for a normal environment.
     """
 
-    UNSPECIFIED = 'AUDIO_MODE_UNSPECIFIED'
+    Unspecified = 'AUDIO_MODE_UNSPECIFIED'
     """
     Audio Mode is not specified.
     """
@@ -40,6 +40,6 @@ class SoundTouchAudioModes(Enum):
     @staticmethod    
     def ToString(value) -> str:
         """ Returns the enum.value (instead of classname.value) as a string. """
-        if isinstance(value, SoundTouchAudioModes):
+        if isinstance(value, AudioDspAudioModes):
             return value.value
         return str(value)