+++ bosesoundtouchapi/models/searchresponse.py
@@ -133,6 +133,8 @@ class SearchResponse:
                 request body; otherwise, False to return all attributes.
         """
         elm = Element('items')
+        elm.set('totalItems', str(self._TotalItems))
+        elm.set('itemCount', str(self.ItemCount))
         
         item:NavigateItem
         for item in self._Items: